Add coldfront container and installation files.
This commit adds a Coldfront container to docker-compose. Coldfront is
installed as it would be in a production setup, i.e. behind nginx with
ssl, coldfront runs on unix socket via gunicorn, and connects to mysql
database. We use the django_pam authentication backend so we can use the
local system user accounts and passwords that are created in the
container. This allows us to login with the same user credentials we use
to submit slurm jobs etc.