Move the editing of the configuration files into the correct place.
The scripts are structured to split out different categories of setup. The
install script is split into two sections: install of packages and
dependencies and update of mandatory O/S configuration.
We don't want the install script to contain tutorial-specifc configuration since we
discuss this at a different point.

Instead the tutorial-specific configuration happens when we discuss the
entry_point script.