Added some error cases so that the script would work
Ran into an issue where pexpect would hang on this `expect` statement and it
would throw a timeout exception. Read through some of the documentation and
noticed that they had `EOF` and `TIMEOUT` explicitly added to their array of
things to look for, so I added them in and the script successfully ran.