Mount slurm state directories as volumes.
Fixes #119. After stopping/starting containers the `hpcadmin` user could
no longer submit jobs without specifying a slurm account, even though
the slurm association had a DefaultAccount set for that user. This
appears to be due to the slurm state directories not persisting after a
restart. This commit adds volumes for slurmd, slurmctld, and slurmdbd
for the slurm state directory.