Revert incorrect rpm install from #153
The pull request #153 added the appkernels RPM to the build. This was never
in the original tutorial and cannot be simply added without also adding the configuration
files, installing the AKRR and also updating the setup script.

Adding the RPM without the corresponding other pacakges and configuration results
in a broken XDMoD portal.